pyproject.toml
[tool.poetry]
name = "larapida-on-django"
version = "0.1.0"
description = "La Rapida website written in python with django"
authors = ["Davide Di Criscito <davide.dicriscito@gmail.com>"]

[tool.isort]
profile = "black"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
django-polymorphic = "^3.1.0"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^21.1"
django-rest-polymorphic = "^0.1.9"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
